% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildAstar}
\alias{buildAstar}
\title{buildAstar}
\usage{
buildAstar(foinew, ldfnew, grainsize, debug)
}
\arguments{
\item{foinew}{numeric \code{\link[base]{matrix}} representing the file of interest}

\item{ldfnew}{numeric \code{\link[base]{matrix}} representing the linking data file}

\item{grainsize}{integer determining minimum grain size for parallisation}

\item{debug}{Boolean indicating whether to output additional debugging information}
}
\description{
Builds the A* matrix
}
\details{
\code{buildAstar} takes a matrix representing the file of interest and 
a matrix representing the linking data file and creates a matrix that 
can then be used to generating linking scores. Reporting frequency as this
occurs can be specified via the nreport option. This is implemented in C++
to provide a speed increase over implementing it directly in the R equivalent.
}
